refactor chisquare expectation into Dataset
