+++ great_expectations/dataset/sparkdf_dataset.py
@@ -459,63 +459,3 @@ class SparkDFDataset(MetaSparkDFDataset):
             raise NotImplementedError('Not implemented for continuous distributions')
 
         return return_obj
-
-    @DocInherit
-    @MetaSparkDFDataset.column_aggregate_expectation
-    def expect_column_chisquare_test_p_value_to_be_greater_than(
-        self,
-        column,
-        partition_object=None,
-        p=0.05,
-        tail_weight_holdout=0,
-        result_format=None,
-        include_config=False,
-        catch_exceptions=None,
-        meta=None,
-    ):
-        if not is_valid_categorical_partition_object(partition_object):
-            raise ValueError("Invalid partition object.")
-        col_name = column.columns[0]
-        observed_frequencies = self.column_value_counts[col_name]
-        # Convert to Series object to allow joining on index values
-        expected_column = pd.Series(
-            partition_object['weights'], index=partition_object['values'], name='expected'
-        ) * self.column_nonnull_counts[col_name]
-        # Join along the indices to allow proper comparison of both types of possible missing values
-        # test_df = pd.concat([expected_column, observed_frequencies], axis=1, sort=True) # Sort parameter not available before pandas 0.23.0
-        test_df = pd.concat([expected_column, observed_frequencies], axis=1)
-
-        na_counts = test_df.isnull().sum()
-
-        # Handle NaN: if we expected something that's not there, it's just not there.
-        test_df[col_name] = test_df[col_name].fillna(0)
-        # Handle NaN: if something's there that was not expected, substitute the relevant value for tail_weight_holdout
-        if na_counts['expected'] > 0:
-            # Scale existing expected values
-            test_df['expected'] = test_df['expected'] * \
-                (1 - tail_weight_holdout)
-            # Fill NAs with holdout.
-            test_df['expected'] = test_df['expected'].fillna(
-                len(column) * (tail_weight_holdout / na_counts['expected']))
-
-        test_result = stats.chisquare(
-            test_df[col_name], test_df['expected'])[1]
-
-        return_obj = {
-            "success": test_result > p,
-            "result": {
-                "observed_value": test_result,
-                "details": {
-                    "observed_partition": {
-                        "values": test_df.index.tolist(),
-                        "weights": test_df[col_name].tolist()
-                    },
-                    "expected_partition": {
-                        "values": test_df.index.tolist(),
-                        "weights": test_df['expected'].tolist()
-                    }
-                }
-            }
-        }
-
-        return return_obj
